Refactor: Move API integration logic to movieApi.ts
